svg/path/property/getComputedStyle.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><metadata><h:link href="https://svgwg.org/svg2-draft/paths.html#TheDProperty" rel="help"/><h:meta content="d is a property" name="assert"/></metadata><style>.p3{d:path(&apos;M 10 3 H 30&apos;)}.g5{d:path(&apos;m 10 5 h 40&apos;)}.p6{d:inherit}</style><g id="g0"><path id="p1"/><path id="p2" d="M 10 2 H 20"/><path id="p3" class="p3"/><path id="p4" style="d:path('M 10 4 H 40')"/></g><g id="g5" class="g5"><path id="p6" class="p6"/><path id="p7"/></g><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script><![CDATA[
  test(function() {
    var g0 = document.getElementById('g0');
    var p1 = document.getElementById('p1');
    var p2 = document.getElementById('p2');
    var p3 = document.getElementById('p3');
    var p4 = document.getElementById('p4');
    var g5 = document.getElementById('g5');
    var p6 = document.getElementById('p6');
    var p7 = document.getElementById('p7');

    assert_equals(getComputedStyle(g0).d, 'none');
    assert_equals(getComputedStyle(p1).d, 'none');
    assert_equals(getComputedStyle(p2).d, 'path("M 10 2 H 20")');
    assert_equals(getComputedStyle(p3).d, 'path("M 10 3 H 30")');
    assert_equals(getComputedStyle(p4).d, 'path("M 10 4 H 40")');
    assert_equals(getComputedStyle(g5).d, 'path("M 10 5 H 50")');
    assert_equals(getComputedStyle(p6).d, 'path("M 10 5 H 50")');
    assert_equals(getComputedStyle(p7).d, 'none');
  });
  ]]></script></svg>
